YUK-13619 Refactor CTINEXUS class (tcp stream socket connection) YUKRV-7524 - change stream connection error types size_t - replace CtiPortShare ctor _shutdownEvent to throw std::exception error - change PORTQUE DeQueue error type to SOCKWRITE of stream socket write failes.. , updated portField check - return unused extern variable inside rte_expresscom and rte_versacom

+++ yukon-server/COMMON/include/streamConnection.h
@@ -39,10 +39,10 @@ public:
     virtual ~StreamConnection() {};
 
     // The pure virtuals below make this an abstract class
-    virtual bool isValid () const = 0;
-    virtual int  write   (void *buf, int len, const Chrono& timeout) = 0;
-    virtual int  read    (void *buf, int len, const Chrono& timeout, const HANDLE *hAbort) = 0;
-    virtual int  peek    (void *buf, int len) = 0;
+    virtual bool   isValid () const = 0;
+    virtual size_t write   (void *buf, int len, const Chrono& timeout) = 0;
+    virtual size_t read    (void *buf, int len, const Chrono& timeout, const HANDLE *hAbort) = 0;
+    virtual size_t peek    (void *buf, int len) = 0;
 
     // Text Description of connection
     Immutable<std::string> Name;
